+++ Changelog.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.10.0-beta.3
+### Changed
+ - All changes from 4.9.7 - 4.9.9
+ 
+### Fixed
+ - Issues with logic messing up form builder.
+ - EditGrid problems where data would get in a detached state.
+ - Fixed unstranslated component label in Webform
+ - Fixed calculated value doesn`t work when editing data with manual override in true position
+
 ## 4.10.0-beta.2
 ### Fixed
  - Fixed display of selected values in table view
@@ -23,6 +33,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
  - Current submission to be set after draft save.
+
+## 4.9.9
+### Fixed
+ - Issue where checkboxes configured as Radio input would still add checkbox key to data.
+
+### Added
+ - An xs column size.
+ 
+### Changed
+ - Upgrade i18next@19.3.4
+
+## 4.9.8
+### Fixed
+ - Issues where the setValue of nested components would not return the correct changed status.
+ - Issues where read only forms would not apply conditions correctly.
+
+## 4.9.7
+### Fixed
+ - UIP-159: Fixing the UUID import.
+ - T826: Fix nested forms never submitting.
 
 ## 4.9.6
 ### Fixed